Implement 'CodeDropwdown.disabled' traitlet.
To make it possible to disable (freeze) the widget.

Note: The "Setup another code" link is not disabled, because it seems
unexpected that a link would be toggled and it is not strictly necessary
to disable the ability to setup a new code even if the state of the
widget is not supposed to change.